% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectionR6.R
\name{Collection}
\alias{Collection}
\title{R6 object for collections of items}
\description{
Keep a list of items, methods to remove, add, remove/add based on a checkbox-type input,
history, etc.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Collection$new()}}
\item \href{#method-add}{\code{Collection$add()}}
\item \href{#method-delete}{\code{Collection$delete()}}
\item \href{#method-checkbox_adjust}{\code{Collection$checkbox_adjust()}}
\item \href{#method-add_or_delete}{\code{Collection$add_or_delete()}}
\item \href{#method-clear}{\code{Collection$clear()}}
\item \href{#method-is_member}{\code{Collection$is_member()}}
\item \href{#method-members_df}{\code{Collection$members_df()}}
\item \href{#method-n}{\code{Collection$n()}}
\item \href{#method-clone}{\code{Collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$new(key_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$add(id, duplicate = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$delete(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkbox_adjust"></a>}}
\if{latex}{\out{\hypertarget{method-checkbox_adjust}{}}}
\subsection{Method \code{checkbox_adjust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$checkbox_adjust(id, chk, toast = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_or_delete"></a>}}
\if{latex}{\out{\hypertarget{method-add_or_delete}{}}}
\subsection{Method \code{add_or_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$add_or_delete(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_member"></a>}}
\if{latex}{\out{\hypertarget{method-is_member}{}}}
\subsection{Method \code{is_member()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$is_member(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-members_df"></a>}}
\if{latex}{\out{\hypertarget{method-members_df}{}}}
\subsection{Method \code{members_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$members_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n"></a>}}
\if{latex}{\out{\hypertarget{method-n}{}}}
\subsection{Method \code{n()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$n()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
