% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tooltips.R
\name{format_naam_tooltip}
\alias{format_naam_tooltip}
\alias{format_adres_tooltip}
\title{Naam met tooltip (overleden datum), adres met tooltip (laatste adres in Ede)}
\usage{
format_naam_tooltip(naam, overleden)

format_adres_tooltip(
  vwsdatuminschrijving,
  vwsgemeentevaninschrijvingomschrijving,
  straatnaam,
  huisnummer,
  huisletter,
  huisnummertoevoeging,
  postcode,
  woonplaats
)
}
\arguments{
\item{naam}{Naam}

\item{overleden}{Date}
}
\description{
Naam met tooltip (overleden datum), adres met tooltip (laatste adres in Ede)
}
