% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adres_depseudo.R
\name{get_adres_depseudo}
\alias{get_adres_depseudo}
\title{Get all persons at an address (de-anonymized)}
\usage{
get_adres_depseudo(adres, database_object)
}
\arguments{
\item{adres}{is a *reactive* **list** with 'postcode', 'huisnummer', 'huisletter', 'huisnummertoevoeging',
without NA (must be replaced with '')!}
}
\description{
Get all persons at an address (de-anonymized)
}
