% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family_depseudo.R
\name{get_family_depseudo_old}
\alias{get_family_depseudo_old}
\title{Lookup family, and de-anonymize if possible.}
\usage{
get_family_depseudo_old(id_in, database_object)
}
\arguments{
\item{id_in}{A reactive vector with a pseudo-BSN.}

\item{database_object}{An instance of the pseudoData R6 object (includes database connection).}
}
\description{
Get family for a person, and immediately de-anonymize if possible. 
Returns a reactive with the family data (including anonymous data), joined with the 
de-anonymized data.
}
