% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izmSearchModule.R
\name{izmSearchUI}
\alias{izmSearchUI}
\alias{izmSearchModule}
\title{De IZM als standalone shiny module}
\usage{
izmSearchUI(
  id,
  reset_button_class = "btn-secondary",
  reset_button_icon = "refresh",
  design = c("shinydashboard", "softui")
)

izmSearchModule(input, output, session)
}
\arguments{
\item{id}{**Moet zijn** : "izm", anders werkt de onderliggende JS code niet.}

\item{reset_button_class}{CSS Class to apply to the reset button (default: "btn-secondary")}

\item{reset_button_icon}{Icon name to apply to the reset button (default: "refresh")}

\item{input}{Shiny input, dont use.}

\item{output}{Shiny input, dont use.}

\item{session}{Shiny input, dont use.}
}
\description{
De Integrale ZoekMachine als module (UI, Server).
}
