% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izmSearchModule.R
\name{izmSearchUI}
\alias{izmSearchUI}
\alias{izmSearchModule}
\title{De IZM als standalone shiny module}
\usage{
izmSearchUI(id)

izmSearchModule(input, output, session)
}
\arguments{
\item{id}{**Moet zijn** : "izm", anders werkt de onderliggende JS code niet.}

\item{input}{Shiny input, dont use.}

\item{output}{Shiny input, dont use.}

\item{session}{Shiny input, dont use.}
}
\description{
De Integrale ZoekMachine als module (UI, Server).
}
