% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoData.R
\name{pseudoData}
\alias{pseudoData}
\title{R6 class for connection to anonymized database.}
\description{
R6 class for connection to anonymized database.

R6 class for connection to anonymized database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pseudoData$new()}}
\item \href{#method-close}{\code{pseudoData$close()}}
\item \href{#method-query}{\code{pseudoData$query()}}
\item \href{#method-listTables}{\code{pseudoData$listTables()}}
\item \href{#method-replace_na_char}{\code{pseudoData$replace_na_char()}}
\item \href{#method-get_person_brp}{\code{pseudoData$get_person_brp()}}
\item \href{#method-set_relation}{\code{pseudoData$set_relation()}}
\item \href{#method-set_parent_relation}{\code{pseudoData$set_parent_relation()}}
\item \href{#method-set_kind_relation}{\code{pseudoData$set_kind_relation()}}
\item \href{#method-get_huwelijk}{\code{pseudoData$get_huwelijk()}}
\item \href{#method-get_kinderen}{\code{pseudoData$get_kinderen()}}
\item \href{#method-get_family}{\code{pseudoData$get_family()}}
\item \href{#method-get_all_bronnen}{\code{pseudoData$get_all_bronnen()}}
\item \href{#method-get_suite}{\code{pseudoData$get_suite()}}
\item \href{#method-get_menscentraal}{\code{pseudoData$get_menscentraal()}}
\item \href{#method-get_openwave}{\code{pseudoData$get_openwave()}}
\item \href{#method-get_carel}{\code{pseudoData$get_carel()}}
\item \href{#method-get_allegro}{\code{pseudoData$get_allegro()}}
\item \href{#method-get_verhuizingen}{\code{pseudoData$get_verhuizingen()}}
\item \href{#method-get_brp_verh_hst}{\code{pseudoData$get_brp_verh_hst()}}
\item \href{#method-get_family_depseudo}{\code{pseudoData$get_family_depseudo()}}
\item \href{#method-get_verhuizingen_depseudo}{\code{pseudoData$get_verhuizingen_depseudo()}}
\item \href{#method-get_adres_depseudo}{\code{pseudoData$get_adres_depseudo()}}
\item \href{#method-meldingConstructor}{\code{pseudoData$meldingConstructor()}}
\item \href{#method-clone}{\code{pseudoData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$new(
  config_file = getOption("shintobag_conf", "conf/config.yml"),
  schema = "",
  filename = NULL,
  pool = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Not used at the moment (data in SQLite).}

\item{\code{schema}}{Not used at the moment.}

\item{\code{filename}}{Full path to the SQLite database.}

\item{\code{pool}}{If TRUE, uses pool to connect.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$query(txt, glue = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listTables"></a>}}
\if{latex}{\out{\hypertarget{method-listTables}{}}}
\subsection{Method \code{listTables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$listTables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_na_char"></a>}}
\if{latex}{\out{\hypertarget{method-replace_na_char}{}}}
\subsection{Method \code{replace_na_char()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$replace_na_char(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_person_brp"></a>}}
\if{latex}{\out{\hypertarget{method-get_person_brp}{}}}
\subsection{Method \code{get_person_brp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_person_brp(
  pseudo_id = NULL,
  what = c("bsn", "anr", "adres"),
  adres = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_relation"></a>}}
\if{latex}{\out{\hypertarget{method-set_relation}{}}}
\subsection{Method \code{set_relation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$set_relation(data, relation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_parent_relation"></a>}}
\if{latex}{\out{\hypertarget{method-set_parent_relation}{}}}
\subsection{Method \code{set_parent_relation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$set_parent_relation(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_kind_relation"></a>}}
\if{latex}{\out{\hypertarget{method-set_kind_relation}{}}}
\subsection{Method \code{set_kind_relation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$set_kind_relation(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_huwelijk"></a>}}
\if{latex}{\out{\hypertarget{method-get_huwelijk}{}}}
\subsection{Method \code{get_huwelijk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_huwelijk(pseudo_id, what = "bsn")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_kinderen"></a>}}
\if{latex}{\out{\hypertarget{method-get_kinderen}{}}}
\subsection{Method \code{get_kinderen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_kinderen(pseudo_id, what = "bsn")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_family"></a>}}
\if{latex}{\out{\hypertarget{method-get_family}{}}}
\subsection{Method \code{get_family()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_family(pseudo_id, what = c("bsn", "anr"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_bronnen"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_bronnen}{}}}
\subsection{Method \code{get_all_bronnen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_all_bronnen(pseudo_bsn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_suite"></a>}}
\if{latex}{\out{\hypertarget{method-get_suite}{}}}
\subsection{Method \code{get_suite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_suite(pseudo_bsn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_menscentraal"></a>}}
\if{latex}{\out{\hypertarget{method-get_menscentraal}{}}}
\subsection{Method \code{get_menscentraal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_menscentraal(pseudo_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_openwave"></a>}}
\if{latex}{\out{\hypertarget{method-get_openwave}{}}}
\subsection{Method \code{get_openwave()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_openwave(pseudo_id, what = "bsn_nummer")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_carel"></a>}}
\if{latex}{\out{\hypertarget{method-get_carel}{}}}
\subsection{Method \code{get_carel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_carel(pseudo_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_allegro"></a>}}
\if{latex}{\out{\hypertarget{method-get_allegro}{}}}
\subsection{Method \code{get_allegro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_allegro(pseudo_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_verhuizingen"></a>}}
\if{latex}{\out{\hypertarget{method-get_verhuizingen}{}}}
\subsection{Method \code{get_verhuizingen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_verhuizingen(pseudo_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_brp_verh_hst"></a>}}
\if{latex}{\out{\hypertarget{method-get_brp_verh_hst}{}}}
\subsection{Method \code{get_brp_verh_hst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_brp_verh_hst(pseudo_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_family_depseudo"></a>}}
\if{latex}{\out{\hypertarget{method-get_family_depseudo}{}}}
\subsection{Method \code{get_family_depseudo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_family_depseudo(id_in)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_verhuizingen_depseudo"></a>}}
\if{latex}{\out{\hypertarget{method-get_verhuizingen_depseudo}{}}}
\subsection{Method \code{get_verhuizingen_depseudo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_verhuizingen_depseudo(id_in)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_adres_depseudo"></a>}}
\if{latex}{\out{\hypertarget{method-get_adres_depseudo}{}}}
\subsection{Method \code{get_adres_depseudo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$get_adres_depseudo(adres)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meldingConstructor"></a>}}
\if{latex}{\out{\hypertarget{method-meldingConstructor}{}}}
\subsection{Method \code{meldingConstructor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$meldingConstructor(days_ago = 42)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
