% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restCallModule.R
\name{restCallModule}
\alias{restCallModule}
\alias{restAddressCallModule}
\title{Wrapper around imzsearch API}
\usage{
restCallModule(
  input,
  output,
  session,
  what = c("depseudo", "lookup", "dev"),
  pseudo_ids = reactive(NULL),
  parse_result = TRUE
)

restAddressCallModule(input, output, session, adres = reactive(NULL))
}
\arguments{
\item{input}{Shiny module, don't use.}

\item{output}{Shiny module, don't use.}

\item{session}{Shiny module, don't use.}

\item{what}{Either 'depseudo', 'lookup', 'dev'}

\item{pseudo_ids}{A reactive vector with pseudo BSNs to send to API.}
}
\description{
A shiny module function (server only) that returns a reactive with parsed results
from either the 'depseudo' or the 'lookup' endpoint. For the 'dev' endpoint, use \code{\link{restAddressCallModule}}.
}
